+++ src/app/basic-examples/component-io.tsx
@@ -9,13 +9,11 @@ interface OptionButtonProps {
 }
 
 const OptionButton: FC<OptionButtonProps> = ({ option, setOption, active }) => {
-  const handleClick = () => setOption(option);
-
   const activeClassName = active.pipe(
     map(active => active && 'active')
   );
 
-  return <button onClick={handleClick} class={['option-button', activeClassName]}>
+  return <button onClick={() => setOption(option)} class={['option-button', activeClassName]}>
     {option}
   </button>;
 };
@@ -24,14 +22,13 @@ const options = ['Option 1', 'Option 2', 'Option 3'];
 
 export const ComponentOutputsExample = () => {
   const selectedOption = new BehaviorSubject<string>('Option 1');
-  const setOption = (option: string) => selectedOption.next(option);
 
   const optionButtons = options.map(option => {
     const active = selectedOption.pipe(
       map(selectedOpt => selectedOpt === option),
     );
 
-    return <OptionButton option={option} setOption={setOption} active={active} />;
+    return <OptionButton option={option} setOption={option => selectedOption.next(option)} active={active} />;
   });
 
   return <div>